Make ThingIFAPIBuilder as inner class of ThingIFAPI
