+++ thingif/src/main/java/com/kii/thingif/ThingIFAPI.java
@@ -24,6 +24,7 @@ import com.kii.thingif.gateway.Gateway;
 import com.kii.thingif.gateway.PendingEndNode;
 import com.kii.thingif.internal.http.IoTRestClient;
 import com.kii.thingif.internal.http.IoTRestRequest;
+import com.kii.thingif.internal.utils._Log;
 import com.kii.thingif.query.AggregatedResult;
 import com.kii.thingif.query.Aggregation;
 import com.kii.thingif.query.CountFieldType;
@@ -70,6 +71,195 @@ public class ThingIFAPI implements Parcelable {
 
     private final Map<String, Class<? extends Action>> actionTypes;
     private final Map<String, Class<? extends TargetState>> stateTypes;
+
+    public static class Builder {
+
+        private static final String TAG = Builder.class.getSimpleName();
+        private final @NonNull Context context;
+        private final @NonNull KiiApp app;
+        private final @NonNull Owner owner;
+        private @Nullable Target target;
+        private @Nullable String installationID;
+        private @Nullable String tag;
+
+        private final @NonNull Map<String, Class<? extends Action>> actionTypes;
+        private final @NonNull Map<String, Class<? extends TargetState>> stateTypes;
+
+        private Builder(
+                @Nullable Context context,
+                @NonNull KiiApp app,
+                @NonNull Owner owner,
+                @NonNull Map<String, Class<? extends Action>> actionTypes,
+                @NonNull Map<String, Class<? extends TargetState>> stateTypes
+        ) {
+            this.context = context;
+            this.app = app;
+            this.owner = owner;
+            this.actionTypes = actionTypes;
+            this.stateTypes = stateTypes;
+        }
+
+        /** Instantiate new Builder.
+         * @param context Application context.
+         * @param app Kii Cloud Application.
+         * @param owner Specify who uses the ThingIFAPI.
+         * @param actionTypes Map of alias and action class.
+         * @param stateTypes Map of alias and target state class.
+         * @return Builder instance.
+         */
+        @NonNull
+        public static Builder newBuilder(
+                @NonNull Context context,
+                @NonNull KiiApp app,
+                @NonNull Owner owner,
+                @NonNull Map<String, Class<? extends Action>> actionTypes,
+                @NonNull Map<String, Class<? extends TargetState>> stateTypes) {
+            if (context == null) {
+                throw new IllegalArgumentException("context is null");
+            }
+            if (app == null) {
+                throw new IllegalArgumentException("app is null");
+            }
+            if (owner == null) {
+                throw new IllegalArgumentException("owner is null");
+            }
+            return new Builder(context, app, owner, actionTypes, stateTypes);
+        }
+
+        /** Instantiate new Builder.
+         * @param context Application context.
+         * @param app Kii Cloud Application.
+         * @param owner Specify who uses the ThingIFAPI.
+         * @return Builder instance.
+         */
+        @NonNull
+        public static Builder newBuilder(
+                @NonNull Context context,
+                @NonNull KiiApp app,
+                @NonNull Owner owner) {
+            if (context == null) {
+                throw new IllegalArgumentException("context is null");
+            }
+            if (app == null) {
+                throw new IllegalArgumentException("app is null");
+            }
+            if (owner == null) {
+                throw new IllegalArgumentException("owner is null");
+            }
+            return new Builder(
+                    context,
+                    app,
+                    owner,
+                    new HashMap<String, Class<? extends Action>>(),
+                    new HashMap<String, Class<? extends TargetState>>());
+        }
+
+        /**
+         * Instantiate new Builder without Context.
+         * This method is for internal use only. Do not call it from your application.
+         *
+         * @param app Kii Cloud Application.
+         * @param owner Specify who uses the ThingIFAPI.
+         * @return Builder instance.
+         */
+        @NonNull
+        public static Builder _newBuilder(
+                @NonNull KiiApp app,
+                @NonNull Owner owner,
+                @NonNull Map<String, Class<? extends Action>> actionTypes,
+                @NonNull Map<String, Class<? extends TargetState>> stateTypes) {
+            if (app == null) {
+                throw new IllegalArgumentException("app is null");
+            }
+            if (owner == null) {
+                throw new IllegalArgumentException("owner is null");
+            }
+            return new Builder(null, app, owner, actionTypes, stateTypes);
+        }
+
+        /**
+         * Set target thing to the ThingIFAPI.
+         * @param target target of {@link ThingIFAPI} instance.
+         * @return builder instance for chaining call.
+         */
+        public Builder setTarget(Target target) {
+            this.target = target;
+            return this;
+        }
+
+        /** Set tag to this ThingIFAPI instance.
+         * tag is used to distinguish storage area of instance.
+         * <br>
+         * If the api instance is tagged with same string, It will be overwritten.
+         * <br>
+         * If the api instance is tagged with different string, Different key is used to store the
+         * instance.
+         * <br>
+         * <br>
+         * Please refer to {@link ThingIFAPI#loadFromStoredInstance(Context, String)} as well.
+         * @param tag if null or empty string is passed, it will be ignored.
+         * @return builder instance for chaining call.
+         */
+        @NonNull
+        public Builder setTag(@Nullable String tag) {
+            this.tag = tag;
+            return this;
+        }
+
+        /**
+         * Set InstallationID to the ThingIFAPI.
+         * @param installationID installation id
+         * @return builder instance for chaining call.
+         */
+        public Builder setInstallationID(String installationID) {
+            this.installationID = installationID;
+            return this;
+        }
+
+        /** Instantiate new ThingIFAPI instance.
+         * @return ThingIFAPI instance.
+         * @throws IllegalStateException when schema is not present.
+         */
+        @NonNull
+        public ThingIFAPI build() {
+
+            _Log.d(TAG, MessageFormat.format("Initialize ThingIFAPI AppID={0}, AppKey={1}, BaseUrl={2}", app.getAppID(), app.getAppKey(), app.getBaseUrl()));
+            return new ThingIFAPI(this.context, this.tag, app, this.owner, this.target, this.installationID, this.actionTypes, this.stateTypes);
+        }
+
+        /**
+         * Register list of Action subclasses to specified alias. The registered action classes
+         * will be used for serialization/deserialization the action.
+         * If the same alias already registered, then will be updated
+         * @param alias Alias to register
+         * @param actionClass List of Action subclasses
+         * @return builder instance for chaining call.
+         */
+        @NonNull
+        public Builder registerActions(
+                @NonNull String alias,
+                @NonNull Class<? extends Action> actionClass){
+            this.actionTypes.put(alias, actionClass);
+            return this;
+        }
+
+        /**
+         * Register TargetState to specified alias.
+         * The registered stateClass will be used when deserialization state from server.
+         * If the same alias already registered, then will be updated.
+         * @param alias Alias to register.
+         * @param stateClass Class of TargetState subclass.
+         * @return builder instance for chaining call.
+         */
+        @NonNull
+        public Builder registerTargetState(
+                @NonNull String alias,
+                @NonNull Class<? extends TargetState> stateClass) {
+            this.stateTypes.put(alias, stateClass);
+            return this;
+        }
+
+    }
 
     /**
      * Try to load the instance of ThingIFAPI using stored serialized instance.
@@ -210,7 +400,7 @@ public class ThingIFAPI implements Parcelable {
             @NonNull Map<String, Class<? extends Action>> actionTypes,
             @NonNull Map<String, Class<? extends TargetState>> stateTypes
             ) {
-        // Parameters are checked by ThingIFAPIBuilder
+        // Parameters are checked by Builder
         if (context != null) {
             ThingIFAPI.context = context.getApplicationContext();
         }
